# contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vis Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/plot-runs": {
      "post": {
        "tags": [
          "plot runs"
        ],
        "summary": "Create Plot Run",
        "operationId": "create_plot_run_api_v1_plot_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlotRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotRunAccepted"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The request does not match the API contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plot-runs/{run_id}": {
      "get": {
        "tags": [
          "plot runs"
        ],
        "summary": "Get Plot Run",
        "operationId": "get_plot_run_api_v1_plot_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotRunSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plot-runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "plot runs"
        ],
        "summary": "Cancel Plot Run",
        "operationId": "cancel_plot_run_api_v1_plot_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotRunSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The requested operation is not valid for the current run state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plot-runs/{run_id}/questions/{question_id}/answer": {
      "post": {
        "tags": [
          "plot runs"
        ],
        "summary": "Answer Plot Run Question",
        "operationId": "answer_plot_run_question_api_v1_plot_runs__run_id__questions__question_id__answer_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotRunAccepted"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The requested operation is not valid for the current run state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The request does not match the API contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plot-runs/{run_id}/approvals/{approval_id}": {
      "post": {
        "tags": [
          "plot runs"
        ],
        "summary": "Decide Plot Run Approval",
        "operationId": "decide_plot_run_approval_api_v1_plot_runs__run_id__approvals__approval_id__post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotRunAccepted"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The requested operation is not valid for the current run state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The request does not match the API contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plot-runs/{run_id}/events": {
      "get": {
        "tags": [
          "plot runs"
        ],
        "summary": "Stream Plot Run Events",
        "operationId": "stream_plot_run_events_api_v1_plot_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered Server-Sent Events. Each data field conforms to run-events-v1.schema.json. Last-Event-ID replays later events.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisResult": {
        "properties": {
          "contains_demo_data": {
            "type": "boolean",
            "title": "Contains Demo Data",
            "default": false
          },
          "result_id": {
            "type": "string",
            "title": "Result Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ObjectReference"
            },
            "type": "array",
            "title": "Inputs"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/ObjectDescription"
            },
            "type": "array",
            "title": "Objects"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactReference"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "code_hash": {
            "type": "string",
            "title": "Code Hash"
          },
          "parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Parameters"
          },
          "environment": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Environment"
          },
          "random_seed": {
            "type": "integer",
            "title": "Random Seed"
          },
          "relationships_used": {
            "items": {
              "$ref": "#/components/schemas/ObjectRelationship"
            },
            "type": "array",
            "title": "Relationships Used"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "result_id",
          "project_id",
          "run_id",
          "name",
          "description",
          "created_at",
          "inputs",
          "objects",
          "code_hash",
          "random_seed"
        ],
        "title": "AnalysisResult"
      },
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recoverable": {
            "type": "boolean",
            "title": "Recoverable",
            "default": false
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApiErrorEnvelope": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope"
      },
      "Approval": {
        "properties": {
          "approval_id": {
            "type": "string",
            "title": "Approval Id"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "scientific_effect": {
            "type": "string",
            "title": "Scientific Effect"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "approval_id",
          "operation",
          "summary",
          "scientific_effect"
        ],
        "title": "Approval"
      },
...
      "ApprovalDecisionRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ApprovalDecisionRequest"
      },
...
      "AutoDataScope": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "auto",
            "title": "Mode",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AutoDataScope"
      },
...
      "CreatePlotRunRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "request": {
            "$ref": "#/components/schemas/PlotRequest"
          },
          "data_scope": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AutoDataScope"
              },
              {
                "$ref": "#/components/schemas/SelectedDataScope"
              },
              {
                "$ref": "#/components/schemas/DemoDataScope"
              }
            ],
            "title": "Data Scope",
            "discriminator": {
              "propertyName": "mode",
              "mapping": {
                "auto": "#/components/schemas/AutoDataScope",
                "demo": "#/components/schemas/DemoDataScope",
                "selected": "#/components/schemas/SelectedDataScope"
              }
            }
          },
          "base_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version Id"
          },
          "research_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchPlan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "request"
        ],
        "title": "CreatePlotRunRequest"
      },
...
      "DemoDataScope": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "demo",
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "DemoDataScope",
        "description": "Explicitly select illustrative data instead of project data."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "PlotRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "title": "Text",
            "default": ""
          },
          "reference_image_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Image Ids"
          },
          "generation_mode": {
            "$ref": "#/components/schemas/GenerationMode",
            "default": "auto"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          },
          "gallery_mode": {
            "$ref": "#/components/schemas/GalleryMode",
            "default": "off"
          },
          "gallery_reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gallery Reference Id"
          },
          "controls_mode": {
            "$ref": "#/components/schemas/ControlsMode",
            "default": "hybrid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PlotRequest"
      },
...
      "PlotRunAccepted": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "links": {
            "$ref": "#/components/schemas/RunLinks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "stage",
          "links"
        ],
        "title": "PlotRunAccepted"
      },
      "PlotRunSnapshot": {
        "properties": {
          "analysis_results": {
            "items": {
              "$ref": "#/components/schemas/AnalysisResult"
            },
            "type": "array",
            "title": "Analysis Results"
          },
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlotResultSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunProgress"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Question"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Approval"
              },
              {
                "type": "null"
              }
            ]
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunFailure"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "project_id",
          "status",
          "stage",
          "created_at",
          "updated_at"
        ],
        "title": "PlotRunSnapshot"
      },
...
      "Question": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/QuestionChoice"
            },
            "type": "array",
            "minItems": 2,
            "title": "Choices"
          },
          "allow_free_text": {
            "type": "boolean",
            "title": "Allow Free Text",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question_id",
          "prompt",
          "reason",
          "choices"
        ],
        "title": "Question"
      },
      "QuestionAnswerRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "choice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choice Id"
          },
          "free_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "QuestionAnswerRequest"
      },
...
      "ResearchPlan": {
        "properties": {
          "figure_size": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FigureSize"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explicit output width and height in inches, chosen for this figure's layout and the user's output requirements. Required when render_code is present; null for analysis without a figure. No default canvas size is assumed."
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "title": "Description"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ResearchInput"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Inputs"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/ObjectRelationship"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Relationships"
          },
          "analysis_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40000
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Code"
          },
          "analysis_parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Analysis Parameters"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/ResearchOutput"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Outputs"
          },
          "reuse_result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reuse Result Id"
          },
          "render_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40000
              },
              {
                "type": "null"
              }
            ],
            "title": "Render Code"
          },
          "controls": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RenderNumberControl"
                },
                {
                  "$ref": "#/components/schemas/RenderChoiceControl"
                },
                {
                  "$ref": "#/components/schemas/RenderTextControl"
                },
                {
                  "$ref": "#/components/schemas/RenderBooleanControl"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "boolean": "#/components/schemas/RenderBooleanControl",
                  "choice": "#/components/schemas/RenderChoiceControl",
                  "number": "#/components/schemas/RenderNumberControl",
                  "text": "#/components/schemas/RenderTextControl"
                }
              }
            },
            "type": "array",
            "maxItems": 42,
            "title": "Controls"
          },
          "control_groups": {
            "items": {
              "$ref": "#/components/schemas/ControlGroup"
            },
            "type": "array",
            "title": "Control Groups"
          },
          "random_seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Random Seed",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "title": "ResearchPlan"
      },
...
      "RunLinks": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "events": {
            "type": "string",
            "title": "Events"
          },
          "cancel": {
            "type": "string",
            "title": "Cancel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "events",
          "cancel"
        ],
        "title": "RunLinks"
      },
      "RunProgress": {
        "properties": {
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "progress",
          "message"
        ],
        "title": "RunProgress"
      },
      "RunStage": {
        "type": "string",
        "enum": [
          "received",
          "understanding_intent",
          "selecting_data",
          "profiling_data",
          "planning_transformations",
          "planning_plot",
          "checking_capabilities",
          "generating_code",
          "running_r",
          "validating_plot",
          "committing_version"
        ],
        "title": "RunStage"
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "awaiting_input",
          "awaiting_approval",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus"
      },
...
      "SelectedDataScope": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "selected",
            "title": "Mode"
          },
          "bundle_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Bundle Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "bundle_ids"
        ],
        "title": "SelectedDataScope"
      },
...
    }
  }
}